+++ src/astronomy/orbit.rs
@@ -3,7 +3,7 @@ use std::time::Duration;
 use bevy::prelude::*;
 use bevy_spatial::{AutomaticUpdate, SpatialStructure};
 
-use crate::resources::{game_time::GameTime, state::GameState};
+use crate::resources::{game_time::GameTime, spatial::KDNode, state::GameState};
 
 const ORBITAL_PERIOD_SCALING_FACTOR: f32 = 1.0;
 
@@ -31,8 +31,7 @@ pub struct OrbitPlugin;
 impl Plugin for OrbitPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(
-            // TODO: Replace use of Orbitable in KDTree with KDNode for reusability.
-            AutomaticUpdate::<Orbitable>::new()
+            AutomaticUpdate::<KDNode>::new()
                 .with_spatial_ds(SpatialStructure::KDTree2)
                 .with_frequency(Duration::from_millis(1)),
         );