+++ src/astronomy/planet.rs
@@ -3,6 +3,7 @@ use bevy::prelude::*;
 use crate::{
     effects::animate::{AnimationBundle, AnimationIndices, AnimationTimer},
     hud::indicator::Indicated,
+    resources::spatial::KDNode,
 };
 
 use super::orbit::{Orbit, Orbitable};
@@ -24,6 +25,7 @@ pub struct PlanetBundle {
     pub orbit: Orbit,
     pub animation_bundle: AnimationBundle,
     pub sprite_sheet_bundle: SpriteSheetBundle,
+    pub kd_node: KDNode,
 }
 impl Default for PlanetBundle {
     fn default() -> Self {
@@ -43,6 +45,7 @@ impl Default for PlanetBundle {
                 timer: AnimationTimer(Timer::from_seconds(0.1, TimerMode::Repeating)),
             },
             sprite_sheet_bundle: SpriteSheetBundle::default(),
+            kd_node: KDNode,
         }
     }
 }