+++ src/resources/mod.rs
@@ -1,4 +1,5 @@
 pub mod assets;
 pub mod camera;
 pub mod game_time;
+pub mod spatial;
 pub mod state;