+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add and use KDNode instead of Orbitable for nearest neighbour mapping
+
 ### Changed
 
 - Separate nav and speedometer UI features into own modules